pyproject.toml
[tool.poetry]
name = "analytics"
version = "0.1.0"
description = ""
authors = ["NistsiukAlexandr"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.109.0"}
aioboto3 = "^12.2.0"
aiokafka = "^0.10.0"
dependency-injector = "^4.41.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
pytest-asyncio = "^0.23.4"
pytest-mock = "^3.12.0"
faker = "^23.1.0"
schemathesis = "^3.25.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
asyncio_mode="auto"
